requirements.txt
Django==1.11.1
djangorestframework==3.6.3
django-jsonfield==1.0.1
django-cors-headers==2.1.0